Rename existing SLO index template
