+++ x-pack/plugins/observability/server/assets/index_templates/slo_index_template.ts
